Remove Git Extended node and add example node
