+++ package.json
@@ -1,19 +1,19 @@
 {
-  "name": "n8n-nodes-extended-git",
+  "name": "n8n-nodes-extended-template",
   "version": "0.1.1",
-  "description": "Community nodes for n8n including a Git Extended node.",
+  "description": "Template for extended n8n custom nodes including an example node.",
   "keywords": [
     "n8n-community-node-package"
   ],
   "license": "MIT",
-  "homepage": "https://github.com/Xopoko/n8n-nodes-extended-git",
+  "homepage": "https://github.com/your-org/n8n-nodes-extended-template",
   "author": {
     "name": "Xopoko",
     "email": "cyprus.horoko@gmail.com"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/Xopoko/n8n-nodes-extended-git.git"
+    "url": "https://github.com/your-org/n8n-nodes-extended-template.git"
   },
   "engines": {
     "node": ">=20.15"
@@ -35,7 +35,7 @@
     "n8nNodesApiVersion": 1,
     "credentials": [],
     "nodes": [
-      "dist/nodes/GitExtended/GitExtended.node.js"
+      "dist/nodes/Example/Example.node.js"
     ]
   },
   "devDependencies": {
